feat: float time now used instead of int
#363

+++ server/src/server/api/actions.py
@@ -29,7 +29,7 @@ async def post_submission(submission: SubmissionRequest, auth_header: dict, toke
         problem_id=submission.problem_id,
         user_uuid=UUID(jwt_to_data(token, settings.JWT_SECRET_KEY, settings.JWT_ALGORITHM).uuid),
         language=submission.language,
-        timestamp=int(datetime.now().timestamp()),
+        timestamp=float(datetime.now().timestamp()),
         code=submission.code,
     )
